+++ CHANGELOG.md
@@ -1,11 +1,19 @@
-## 0.0.1
+## 1.0.0+2
 
-- Initial publish package.
+- add alpabet to jump specific country
+
+## 1.0.0+1
+
+- add search list
 
 ## 1.0.0
 
 - Clean some unusefull code
 
-## 1.0.0+1
+## 0.0.1
 
-- add search list
+- Initial publish package.
+
+
+
+